+++ CHANGELOG.md
@@ -9,6 +9,10 @@ Guide: https://keepachangelog.com/en/1.0.0/
 <!-- Add changes for active work here -->
 
 - [Discover] Add support for country, proximity, and origin parameters in Discover.Options search parameters. This fixes an issue when using search-along-route to query category results.
+- [Unit Tests] Update and correct tests for iOS 17 using all mocked data.
+- [UI Tests] Update and correct tests for iOS 17 using all mocked data.
+- [Search] Rename `SearchEngine.reverseGeocoding` function to `SearchEngine.reverse`.
+- [Core] Stop reading "MapboxAPIBaseURL" from UserDefaults in `ServiceProvider.createEngine`. (Providing a value in Info.plist is still supported).
 - [Core] Remove Swifter library dependency from MapboxSearch target (only used in Test targets)
 - [Core] Change AbstractSearchEngine.init `supportSBS: Bool = false` parameter to `apiType: ApiType = .SBS`. This changes the default API engine for discover/category and other API requests to SBS. Add ApiType enum to represent non-Autofill and non-PlaceAutocomplete SearchEngine API types.